% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/claims.R
\name{claims_summary}
\alias{claims_summary}
\title{Counts by verdict}
\usage{
claims_summary(claims = read_claims())
}
\arguments{
\item{claims}{as returned by \code{\link[=read_claims]{read_claims()}}.}
}
\description{
Counts by verdict
}
